+++ sdk/storage/azure-storage-blob-cryptography/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Release History
+## Version XX.X.X (XXXX-XX-XX)
+- Added a check in EncryptedBlobClientBuilder to enforce HTTPS for bearer token authentication.
+
 ## Version 12.0.0 (2019-10-31)
 - Removed EncryptedBlobClientBuilder inheritance of BaseBlobClientBuilder